Reduce invocations of ParseTag
ParseTag is comparably expensive: reading the entire tag contents and
applying multiple string matches against it. Reduce invocations of
ParseTag by virtue of making Window an observer on the tag's
file.ObservableEditableBuffer. Helps with #400.
