+++ wind.go
@@ -5,13 +5,11 @@ import (
 	"image"
 	"os"
 	"path/filepath"
-	"strings"
 	"sync"
 
 	"github.com/rjkroege/edwood/draw"
 	"github.com/rjkroege/edwood/file"
 	"github.com/rjkroege/edwood/frame"
-	"github.com/rjkroege/edwood/runes"
 	"github.com/rjkroege/edwood/util"
 )
 
@@ -53,16 +51,21 @@ type Window struct {
 	utflastqid  int    // Qid of last read request (QWbody or QWtag)
 	utflastboff uint64 // Byte offset of last read of body or tag
 	utflastq    int    // Rune offset of last read of body or tag
-	tagsafe     bool
-	tagexpand   bool
-	taglines    int
-	tagtop      image.Rectangle
-	editoutlk   chan bool
+
+	tagfilenameend     int
+	tagfilenamechanged bool
+	tagsetting         bool
+	tagsafe            bool // What is tagsafe for?
+	tagexpand          bool
+	taglines           int
+	tagtop             image.Rectangle
+
+	editoutlk chan bool
 }
 
 var (
-	_ file.BufferObserver    = (*Window)(nil) // Enforce at compile time that Window implements BufferObserver
 	_ file.TagStatusObserver = (*Window)(nil) // Enforce at compile time that Window implements BufferObserver
+	_ file.BufferObserver    = (*Window)(nil) // Enforce at compile time that TagIndex implements BufferObserver
 )
 
 func NewWindow() *Window {
@@ -186,7 +189,7 @@ func (w *Window) DrawButton() {
 }
 
 func (w *Window) delRunePos() int {
-	i := len([]rune(w.ParseTag())) + 2
+	i := w.tagfilenameend + 2
 	if i >= w.tag.Nc() {
 		return -1
 	}
@@ -398,198 +401,34 @@ func (w *Window) Undo(isundo bool) {
 
 	// TODO(rjk): Is this absolutely essential.
 	body.Show(body.q0, body.q1, true)
-
 }
 
 func (w *Window) SetName(name string) {
 	t := &w.body
 	t.file.SetName(name)
-
 }
 
 func (w *Window) Type(t *Text, r rune) {
 	t.Type(r)
 }
 
-func (w *Window) ClearTag() {
-	// w must be committed
-	n := w.tag.Nc()
-	r := make([]rune, n)
-	w.tag.file.Read(0, r)
-	i := len([]rune(w.ParseTag()))
-	for ; i < n; i++ {
-		if r[i] == '|' {
-			break
-		}
-	}
-	if i == n {
-		return
-	}
-	i++
-	w.tag.Delete(i, n, true)
-	w.tag.file.Clean()
-	if w.tag.q0 > i {
-		w.tag.q0 = i
-	}
-	if w.tag.q1 > i {
-		w.tag.q1 = i
-	}
-	w.tag.SetSelect(w.tag.q0, w.tag.q1)
-}
-
-// ParseTag returns the filename in the window tag.
-func (w *Window) ParseTag() string {
-	r := make([]rune, w.tag.Nc())
-	w.tag.file.Read(0, r)
-	tag := string(r)
-
-	// " |" or "\t|" ends left half of tag
-	// If we find " Del Snarf" in the left half of the tag
-	// (before the pipe), that ends the file name.
-	pipe := strings.Index(tag, " |")
-	if i := strings.Index(tag, "\t|"); i >= 0 && (pipe < 0 || i < pipe) {
-		pipe = i
-	}
-	if i := strings.Index(tag, " Del Snarf"); i >= 0 && (pipe < 0 || i < pipe) {
-		return tag[:i]
-	}
-	if i := strings.IndexAny(tag, " \t"); i >= 0 {
-		return tag[:i]
-	}
-	return tag
-}
-
-// ForceSetWindowTag force sets the tag when the tag needs to change
-// without a body modification to trigger the tag update.
-// TODO(rjk): Uses of this method are probably code cleanup opportuniies.
-func (w *Window) ForceSetWindowTag() {
-	if w.col.safe || w.tag.fr.GetFrameFillStatus().Maxlines > 0 {
-		w.setTag1()
-	}
-}
-
-// Use to debug.
-// func (w *Window) setTagDiag(where, when string) {
-// 	log.Println("<<", where, when, w.tag.q0, w.tag.q1, w.tag.DebugString())
-// }
-
-// setTag1 updates the tag contents for a given window w.
-// TODO(rjk): Note the col.safe test... should I do this as part of setTag1()?
-func (w *Window) setTag1() {
-	// w.setTagDiag("setTag1", "before")
-	// defer w.setTagDiag("setTag1","after")
-
-	// TODO(rjk): Figure out if I need this. Presumably this is needed to
-	// make things display correctly when filesystem changes to the tag
-	// happen while the window is collapsed to 0?
-	if !w.col.safe && w.tag.fr.GetFrameFillStatus().Maxlines == 0 {
-		// log.Println("Window.setTag1 early exit")
-		return
-	}
-
-	const (
-		Ldelsnarf = " Del Snarf"
-		Lundo     = " Undo"
-		Lredo     = " Redo"
-		Lget      = " Get"
-		Lput      = " Put"
-		Llook     = " Look"
-		Ledit     = " Edit"
-		Lpipe     = " |"
-	)
-
-	// (flux) The C implemtation does a lot of work to avoid
-	// re-setting the tag text if unchanged.  That's probably not
-	// relevant in the modern world.  We can build a new tag trivially
-	// and put up with the traffic implied for a tag line.
-
-	var sb strings.Builder
-	sb.WriteString(w.body.file.Name())
-	sb.WriteString(Ldelsnarf)
-
-	if w.filemenu {
-		if w.body.needundo || w.body.file.HasUndoableChanges() {
-			sb.WriteString(Lundo)
-		}
-		if w.body.file.HasRedoableChanges() {
-			sb.WriteString(Lredo)
-		}
-		if w.body.file.SaveableAndDirty() {
-			sb.WriteString(Lput)
-		}
-	}
-	// TODO(rjk): What happens if I make a directory into
-	if w.body.file.IsDir() {
-		sb.WriteString(Lget)
-	}
-	oldbarIndex := w.tag.file.IndexRune('|')
-	if oldbarIndex >= 0 {
-		// TODO(rjk): Update for file.Buffer representation.
-		oldsuffix := make([]rune, w.tag.file.Nr()-oldbarIndex)
-		w.tag.file.Read(oldbarIndex, oldsuffix)
-		sb.WriteString(" ")
-		sb.WriteString(string(oldsuffix))
-	} else {
-		sb.WriteString(Lpipe)
-		sb.WriteString(Llook)
-		sb.WriteString(Ledit)
-		sb.WriteString(" ")
-	}
-
-	newtag := []rune(sb.String())
-
-	// replace tag if the new one is different
-	resize := false
-	if !runes.Equal(newtag, []rune(w.tag.file.String())) {
-		resize = true // Might need to resize the tag
-		// try to preserve user selection
-		newbarIndex := runes.IndexRune(newtag, '|') // New always has '|'
-		q0 := w.tag.q0
-		q1 := w.tag.q1
-
-		// These alter the Text's selection values.
-		w.tag.Delete(0, w.tag.Nc(), true)
-		w.tag.Insert(0, newtag, true)
-
-		// Rationalize the selection as best as possible
-		w.tag.q0 = util.Min(q0, w.tag.Nc())
-		w.tag.q1 = util.Min(q1, w.tag.Nc())
-		if oldbarIndex != -1 && q0 > oldbarIndex {
-			bar := newbarIndex - oldbarIndex
-			w.tag.q0 = q0 + bar
-			w.tag.q1 = q1 + bar
-		}
-	}
-	w.tag.file.Clean()
-	n := w.tag.file.Nr()
-	if w.tag.q0 > n {
-		w.tag.q0 = n
-	}
-	if w.tag.q1 > n {
-		w.tag.q1 = n
-	}
-	// TODO(rjk): This may redraw the selection unnecessarily
-	// if we replaced the tag above.
-	w.tag.SetSelect(w.tag.q0, w.tag.q1)
-	w.DrawButton()
-	if resize {
-		w.tagsafe = false
-		w.Resize(w.r, true, true)
-	}
-}
-
 // TODO(rjk): In the future of File's replacement with undo buffer,
-// this method could be renamed to something like "UpdateTag"
+// this method could be renamed to something like "UpdateTag"?
 func (w *Window) Commit(t *Text) {
 	t.Commit()
 	if t.what == Body {
 		return
 	}
-	filename := w.ParseTag()
-	if filename != w.body.file.Name() {
-		global.seq++
-		w.body.file.Mark(global.seq)
-		w.SetName(filename)
+	// TODO(rjk): By virtue of being an observer, we know when this has
+	// changed. No need to extract it here unless its changed.
+	if w.tagfilenamechanged {
+		filename := w.ParseTag()
+		if filename != w.body.file.Name() {
+			global.seq++
+			w.body.file.Mark(global.seq)
+			w.SetName(filename)
+		}
+		w.tagfilenamechanged = false
 	}
 }
 
@@ -612,6 +451,7 @@ func isDir(r string) (bool, error) {
 	return false, nil
 }
 
+// Should include file lookup be built-in? Or provided by a helper?
 func (w *Window) AddIncl(r string) {
 	// Tries to open absolute paths, and if fails, tries
 	// to use dirname instead.
@@ -714,16 +554,9 @@ func (w *Window) ClampAddr() {
 	}
 }
 
-// TODO(rjk): need to watch the tag for inserts / deletes so that I can
-// update the tag index structure.
-func (w *Window) Inserted(q0 int, r []rune) {
-}
-
-func (w *Window) Deleted(q0, q1 int) {
-}
-
 // Watch for filename undo actions.
 func (w *Window) MemoizedUndone(undo bool) {
+	// log.Println("Window.MemoizedUndone")
 }
 
 // TODO(rjk): This should only get invokved when the tag needs to change
